Extracts metacircular & pretty from main file
