+++ src/main/scala/core/Statelesser.scala
@@ -1,4 +1,5 @@
 package statelesser
+package core
 
 import Function.const
 import scalaz._, Scalaz._
@@ -69,127 +70,9 @@ object Statelesser {
 
   def apply[E[_]](implicit ev: Statelesser[E]): Statelesser[E] = ev
 
-  implicit val prettyPrinter = new Statelesser[Const[String, ?]] {
+  implicit val prettyPrinter = new interpreter.PrettyPrinter
 
-    def flVert[S, A, B](
-        u: Const[String, Fold[S, A]],
-        d: Const[String, Fold[A, B]]): Const[String, Fold[S, B]] =
-      Const(s"${u.getConst} > ${d.getConst}")
-
-    def flHori[S, A, B](
-        l: Const[String, Fold[S, A]],
-        r: Const[String, Fold[S, B]]): Const[String, Fold[S, (A, B)]] =
-      Const(s"${l.getConst} * ${r.getConst}")
-
-    def gtVert[S, A, B](
-        u: Const[String, Getter[S, A]],
-        d: Const[String, Getter[A, B]]): Const[String, Getter[S, B]] =
-      Const(s"${u.getConst} > ${d.getConst}")
-
-    def gtHori[S, A, B](
-        l: Const[String, Getter[S, A]],
-        r: Const[String, Getter[S, B]]): Const[String, Getter[S, (A, B)]] =
-      Const(s"(${l.getConst} * ${r.getConst})")
-
-    def aflVert[S, A, B](
-        u: Const[String, AffineFold[S, A]],
-        d: Const[String, AffineFold[A, B]]): Const[String, AffineFold[S, B]] =
-      Const(s"${u.getConst} > ${d.getConst}")
-
-    def aflHori[S, A, B](
-        l: Const[String, AffineFold[S, A]],
-        r: Const[String, AffineFold[S, B]]): Const[String, AffineFold[S, (A, B)]] =
-      Const(s"${l.getConst} * ${r.getConst}")
-
-    def filtered[S](
-        p: Const[String, Getter[S, Boolean]]): Const[String, AffineFold[S, S]] =
-      Const(s"filtered(${p.getConst})")
-
-    def sub: Const[String, Getter[(Int, Int), Int]] =
-      Const("sub")
-
-    def greaterThan: Const[String, Getter[(Int, Int), Boolean]] =
-      Const("greaterThan")
-
-    def first[A, B]: Const[String, Getter[(A, B), A]] =
-      Const("first")
-
-    def second[A, B]: Const[String, Getter[(A, B), B]] =
-      Const("second")
-
-    def likeInt[S](i: Int): Const[String, Getter[S, Int]] =
-      Const(s"likeInt(${i.toString})")
-
-    def likeBool[S](b: Boolean): Const[String, Getter[S, Boolean]] =
-      Const(s"likeBool(${b.toString})")
-
-    def likeStr[S](s: String): Const[String, Getter[S, String]] =
-      Const(s"""likeStr("$s")""")
-
-    def id[S]: Const[String, Getter[S, S]] =
-      Const("id")
-
-    def not: Const[String, Getter[Boolean, Boolean]] =
-      Const("not")
-
-    def gtAsAfl[S, A](
-        gt: Const[String, Getter[S, A]]): Const[String, AffineFold[S, A]] =
-      Const(s"${gt.getConst}.asAffineFold")
-
-    def aflAsFl[S, A](
-        afl: Const[String, AffineFold[S, A]]): Const[String, Fold[S, A]] =
-      Const(s"${afl.getConst}.asFold")
-  }
-
-  implicit def metacircular: Statelesser[Id] = new Statelesser[Id] {
-
-    def flVert[S, A, B](u: Fold[S, A], d: Fold[A, B]): Fold[S, B] =
-      u > d
-
-    def flHori[S, A, B](l: Fold[S, A], r: Fold[S, B]): Fold[S, (A, B)] =
-      l * r
-
-    def gtVert[S, A, B](u: Getter[S, A], d: Getter[A, B]): Getter[S, B] =
-      u > d
-
-    def gtHori[S, A, B](l: Getter[S, A], r: Getter[S, B]): Getter[S, (A, B)] =
-      l * r
-
-    def aflVert[S, A, B](
-        u: AffineFold[S, A], 
-        d: AffineFold[A, B]): AffineFold[S, B] =
-      u > d
-
-    def aflHori[S, A, B](
-        l: AffineFold[S, A],
-        r: AffineFold[S, B]): AffineFold[S, (A, B)] =
-      l * r
-
-    def filtered[S](p: Getter[S, Boolean]): AffineFold[S, S] =
-      AffineFold.filtered(p)
-
-    def sub: Getter[(Int, Int), Int] = Getter.sub
-
-    def greaterThan: Getter[(Int, Int), Boolean] = Getter.gt
-
-    def first[A, B]: Getter[(A, B), A] = Getter.first
-
-    def second[A, B]: Getter[(A, B), B] = Getter.second
-
-    def likeInt[S](i: Int): Getter[S, Int] = Getter.like(i)
-
-    def likeBool[S](b: Boolean): Getter[S, Boolean] = Getter.like(b)
-
-    def likeStr[S](s: String): Getter[S, String] = Getter.like(s)
-
-    def id[S]: Getter[S, S] = Getter.id
-
-    def not: Getter[Boolean, Boolean] = Getter.not
-
-    def gtAsAfl[S, A](gt: Getter[S, A]): AffineFold[S, A] = gt.asAffineFold
-
-    def aflAsFl[S, A](afl: AffineFold[S, A]): Fold[S, A] = afl.asFold
-  }
+  implicit def metacircular = new interpreter.Metacircular
 
   implicit def tsemantic: Statelesser[Semantic] = new Statelesser[Semantic] {
 
@@ -339,7 +222,7 @@ object Statelesser {
       val (t, rws) = done.vars
         .groupBy(_._2)
         .map { case (k, v) => (k, v.keys.toList) }
-        .foldLeft((Map.empty[Symbol, Value], Set.empty[(String, String)])) {
+        .foldLeft((Map.empty[Symbol, statelesser.Value], Set.empty[(String, String)])) {
           case ((m, rws), (v, k :: ks)) =>
             (m + (k -> v), rws ++ ks.map((_, k)))
         }
